backend/taste-profile-builder-python/requirements.txt
pandas
scikit-learn
kafka-python